+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changes
 
+## 0.2.4
+* Prints Rust stack trace for PC platforms and for JNI errors other than Java exceptions.
+* Gets the Android context object from `ActivityThread.currentActivityThread().getApplication()` and prints a warning message in case of `ndk_context::android_context().context()` is null: functionalities related to the UI or `Activity` will not work, and the glue crate should be checked.
+
 ## 0.2.3
 * Fixed the bug of panicking during building instead of using the prebuilt dex fallback when some environment variables required by `android-build` are missing.
 * Removed `android-build` build dependency, because of its complicated configuration.